fix(authz): wire ABAC PDP into all auditquery-serving assemblies — F1 fanout [#1348]
CI integration shards (corebundle + tests) went red after the F1 handler fix:
making an empty actorId a permissioned read means EVERY assembly serving
auditquery must wire the PDP, but only cmd/corebundle did. ssobff (admin audit
query → 403) and the slice-level audit_query_projection e2e (empty actorId, no
authorizer → 403) broke; corebundlestarter was latently broken (no failing test).

Fix — share the wiring, wire all roots:
- Relocate the lazy-authorizer + provider-discovery from cmd/corebundle (package
  main, unreusable) to runtime/bootstrap.PrimaryAuthorizerOption(cells) — one
  battle-tested path (lazy resolve + F8 startup fail-fast + nil fail-closed)
  instead of a copy per composition root. Tests move with it.
- cmd/corebundle/run.go + authz_pdp_integration_test.go: call
  bootstrap.PrimaryAuthorizerOption.
- examples/ssobff + examples/corebundlestarter: wire it too (both include
  accesscore, both serve auditquery).
- tests/integration/audit_query_projection_test.go: inject an allow Authorizer
  (mirrors production) so the masking e2e reaches the data layer it tests.

Verified locally (the gap that let round-1 through): tests/integration
TestAuditQueryColumnMaskE2E, cmd/corebundle integration shard (e2e + auth +
setup), examples/ssobff walkthrough (PG testcontainers, admin audit → 200 via
baseline), examples/corebundlestarter, runtime/bootstrap relocated tests — all
green. Build all tags/modules, lint 0, gofumpt clean, archtest 0 new failures.

Refs #1348
