+++ cmd/corebundle/run.go
@@ -102,7 +102,7 @@ func runCorebundle(ctx context.Context, assemblyID string, assemblyCellIDs []str
 		if rtErr != nil {
 			return nil, fmt.Errorf("default runtime options: %w", rtErr)
 		}
-		authzOpt, authzErr := primaryAuthorizerOption(cells)
+		authzOpt, authzErr := bootstrap.PrimaryAuthorizerOption(cells)
 		if authzErr != nil {
 			return nil, fmt.Errorf("primary authorizer wiring: %w", authzErr)
 		}
